pyproject.toml
[project]
name = "maestro"
version = "0.1.0"
description = "CLI Toolkits for Project Maestro (use `uv run *`)"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.3",
    "django>=6.0.4",
    "django-celery-results>=2.6.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "redis>=7.4.0",
]
